requirements.txt
requests>=2.28.0
pyaudio>=0.2.11
SpeechRecognition>=3.8.1
pyttsx3>=2.90
PyPDF2>=2.0.0
python-docx>=0.8.11